Fixed error handling in options.quote. Updated package version to be 0.2.4.
