+++ test/unit/sdk.pricing.ts
@@ -59,6 +59,8 @@ describe('PricingAPI', function (this: any) {
 		'0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80'
 	const emitWarningStub = sinon.stub(process, 'emitWarning')
 
+	this.timeout(10000)
+
 	before(async () => {
 		premia = await Premia.initialize({
 			privateKey: privateKey,
